requirements.txt
Flask==3.0.0
Flask-Babel==4.0.0
requests==2.31.0
langdetect==1.0.9
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0